+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # upcoming
 
 * Register each directive in a separate module to allow for importing only selected directives
-* Merge `templates.js` dist file into `scripts.js`
+* BREAKING CHANGE: Merge `templates.js` and `scripts.js` into `angular-input-fields.js` and `angular-input-fields.min.js`
 * BREAKING CHANGE: directive templates namespaced to `aif`, each moved to a separate folder
 
 # 7.0.1