+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# upcoming
+# 8.0.0
 
 * Register each directive in a separate module to allow for importing only selected directives
 * BREAKING CHANGE: Merge `templates.js` and `scripts.js` into `angular-input-fields.js` and `angular-input-fields.min.js`